backend/requirements.txt
Django
djangorestframework
django-cors-headers
jwt
Pillow
firebase-admin